Basic AI (nothing much), 8-direction Sprites, Attack Updates
Added the basic state-machine and did some code refactors.
Added the 8-direction sprites, so you can vaguely tell what direction the sprites are looking in.
Fixed up some of the stagger code (for that knockback effect) and did refactors.

TODO: implement the AI to do stuff that I want it to. It's not correct yet.