Added Crying Miku for Corpse Sprite, Fixed Some Clamping Issues and Ground Detection Problem
Fixed the issue with the torso hitboxes going below the player feet, causing them to slide down instead of staying still.
Tweaked the player camera position.
Added a sprite for a Crying Miku when you kill her. TODO: Implement it.
Turned off clamping on the ground due to the issues that arise from it in the game.

TODO: Fix the player slide now when they fall off surfaces. Since there is no clamping, they quickly go out of slide as they are not stuck to the ground anymore.